Update pal from commit e1d28c51
* Drop support for gfx8-9
* 128BPP fastclear support
* Fix graphics pipeline library regressions
* Sample coordinates of Fmask surface should be the same as origin image's
* Fix test fail about dEQP-VK.query_pool.statistics_query.*_cq
* Fixup for likely/unlikely, redundant
* Add Local heap gpu memory management for GpaSession
* Use ShaderLibrary owned code object data
* Split ListDir into CountFilesInDir and GetFileNamesInDir
* Adds TraceError RDF chunk
* Force configures the Display DCC offset as displayDccOffset[MaxNumPlanes] when sharedMetadata is used
* Fixup for add SQTT token mask, redundant
* Add a setting to change engine preference for image copies
* Optimize MSAA clears.
* Refactor GetGenericGpuSymbol
* GFX10: AC01 clears can use R9G9B9E5 format
* Discard usage of GpuHeapGartCacheable as a backup heap
* Bump CMake minimum version to 3.21
* Update devdriver
* Refactor ComputeShaderLibrary function list creation
* Add Util::AtomicCompareExchangePointer
* Fix and refactor Continuation stack computation.
* Update Navi31 numSdpInterfaces to 24
* Fix GpuHeapAccess does not handle TMZ heaps properly
* Fix perfcounter failures when measuring a large number of UMCCH counters
* InterpolatorCount is invalid for prerasterOutputSemantic
* Fix warnings that clang 16 complained about
* Fix memory growth in GpaSession
* Remove SQ_NON_EVENT from nested postamble
* Use `[[likely]]` and `[[unlikely]]` instead of `__builtin_expect()`
* Use RingSizes from cmdBuffer.
* VK_EXT_primitives_generated_query - Driver Implementation
* Make some asserts debug only
* Add options SurfaceCaptureFrameStart/FrameCount/FilenameHashType for surface capture
* Add panel setting to control T#/Target.COLOR_TRANSFORM
* Update Coding Standard doc to explicitly allow attributes
* Add GetGpuInfo lookup helpers for IDs and info
* [VK_EXT_nested_command_buffer] Check that the command buffer is not nested when setting allowIb2Launch
* Added build option PAL_BUILD_RDF with default ON
* Update the comment that offset between drawRing and controBuffer must be 256 from pm4 spec
* Add new SQTT token mask setting

+++ src/core/gpuMemory.h
@@ -197,7 +197,7 @@ public:
         const Device&              device,
         const GpuMemoryCreateInfo& createInfo,
         GpuHeap*                   pOutHeaps,
-        size_t*                    pOutHeapCount);
+        uint32*                    pOutHeapCount);
     static Result ValidateCreateInfo(const Device* pDevice, const GpuMemoryCreateInfo& createInfo);
     static Result ValidatePinInfo(const Device* pDevice, const PinnedGpuMemoryCreateInfo& createInfo);
     static Result ValidateSvmInfo(const Device* pDevice, const SvmGpuMemoryCreateInfo& createInfo);
@@ -362,7 +362,7 @@ protected:
 
     Device*const   m_pDevice;
     VaPartition    m_vaPartition;
-    size_t         m_heapCount;
+    uint32         m_heapCount;
     GpuHeap        m_heaps[GpuHeapCount];
 
     // Priority of the memory allocation: serves as a hint to the operating system of how important it is to keep
